Move tile_matmul into tile_matmul.h
Pure refactor: extract the partitioned_gemm helpers, scalar_matmul, the
tile_matmul / tile_matmul_acc entry templates, and adj_tile_matmul out
of tile.h into a dedicated header. Mirrors the per-op-header layout
already used for tile_reduce.h and tile_radix_sort.h. Add an in-file
performance note documenting that the cooperative scalar GEMM path is a
deliberate option for small tiles where libmathdx LTO setup overhead
doesn't pay back, not solely a build-availability fallback.

No behaviour change.
